57 - number input allowed to float

+++ utils/inputs.py
@@ -1,7 +1,7 @@
 def num_input():
     try:
         usr_input = input()
-        usr_input = int(usr_input)
+        usr_input = float(usr_input)
     except ValueError:
         print("That is not a number. \nPlease enter a Number.")
         return None
